+++ python/lvmecp/hvac.py
@@ -25,7 +25,7 @@ class HVACController(PLCModule):
 
         self.status: dict[str, float | bool | None] = {}
 
-    async def _update_internal(self):
+    async def _update_internal(self, **_):
         """Update status."""
 
         self.status = await self.modbus.get_all()