+++ python/lvmecp/safety.py
@@ -19,7 +19,7 @@ class SafetyController(PLCModule[SafetyStatus]):
     """Handles the enclosure safety features."""
 
     flag = SafetyStatus
-    interval = 10.0
+    interval = None
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)